requirements.txt
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
RPi.GPIO>=0.7.1
